feat: adicionado em produto campo de ordem, define posição dentro da categoria
