--- src/constants/urlApi.js
@@ -1,4 +1,0 @@
-// export const urlApi = "http://10.0.110.106:8080/" 
-export const urlApi = "https://cardapiei-api.up.railway.app/"
-export const urlSite = "https://cardapieifacil.com/"
-/* export const urlSite = "http://localhost:3000/" */
