+++ src/services/estabelecimento/schemas/EstabelecimentoFullDetalheResponse.ts
@@ -1,7 +1,7 @@
 import { HorarioAtendimentoSchema } from './../../horarios_atendimento/schemas/HorariosAtendimentoResponse';
 import { ContatoSchema } from "@/services/contato/schemas/ContatoResponse";
 import { LocalicaoSchema } from "@/services/localizacao/schemas/LocalizacaoResponse";
-import { CategoriaSchema } from "@/services/produto/schemas/CategoriaResponse";
+import { CategoriaListSchema } from "@/services/produto/schemas/CategoriaListResponse";
 import { z } from "zod";
 
 export const EstabelecimentoFullDetalheSchema = z.object({
@@ -13,7 +13,7 @@ export const EstabelecimentoFullDetalheSchema = z.object({
     usu_id: z.number(),
 
     //todos os objetos que vem carregado com ele no cardapio
-    Categoria: CategoriaSchema,
+    Categoria: CategoriaListSchema,
     Contatos: ContatoSchema,
     Localizacaos: LocalicaoSchema,
     horario_atendimentos: HorarioAtendimentoSchema
