+++ src/services/produto/schemas/CategoriaListResponse.ts
@@ -1,10 +1,10 @@
 import { ProdutoListSchema } from './ProdutoListResponse';
 import { z } from "zod";
 
-export const CategoriaSchema = z.object({
+export const CategoriaListSchema = z.object({
     id: z.number(),
     nome: z.string(),
     Produtos:ProdutoListSchema.array() 
 }).array()
 
-export type CategoriaResponse = z.infer<typeof CategoriaSchema>;
+export type CategoriaListResponse = z.infer<typeof CategoriaListSchema>;
