+++ src/services/produto/schemas/doDataCategoriasProdutoListRequest.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+
+export const DoDataCategoriasProdutoListSchema = z.object({
+    estId: z.number()
+})
+
+export type DoDataCategoriasProdutoListRequest = z.infer<typeof DoDataCategoriasProdutoListSchema>;
