+++ src/stores/bound.ts
@@ -4,8 +4,9 @@ import { createUserSlice, UserStore } from "./user-store";
 import { createJSONStorage, persist } from "zustand/middleware";
 import { CarrinhoStore, createCarrinhoSlice } from './carrinho/carrinho-slice-store';
 import { ConfiguracaoStore, createConfiguracaoSlice } from './configuracao/configuracao-slice-store';
+import { CardapioAdminStore, createCardapioAdminSlice } from './cardapio-admin/cardapio-admin-slice-store';
 
-type Store = UserStore & EstabelecimentoStore & CarrinhoStore & ConfiguracaoStore
+type Store = UserStore & EstabelecimentoStore & CarrinhoStore & ConfiguracaoStore & CardapioAdminStore
 
 export const useStore = create<Store>()(
     persist(
@@ -13,7 +14,8 @@ export const useStore = create<Store>()(
             ...createUserSlice(...setGet),
             ...createEstabelecimentoSlice(...setGet),
             ...createCarrinhoSlice(...setGet),
-            ...createConfiguracaoSlice(...setGet)
+            ...createConfiguracaoSlice(...setGet),
+            ...createCardapioAdminSlice(...setGet)
         }),
         {
             name: "cardapiei-facil-storage",
